pyproject.toml
[project]
name = "lgctrl"
version = "0.1.0"
description = "App to control/automate my visually-impared Grandma's TV"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiowebostv>=0.7.5",
    "ruff>=0.14.10",
    "wakeonlan>=3.1.0",
    "websockets>=15.0.1",
    "PyYAML>=6.0",
]


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 120  # Same as Black.

target-version = 'py312'  # Assume Python 3.12.

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E722"]

fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

